Renaming AutoComplete to Suggestions
